audio_console_app: add rule to access files on sdcard
If an audio file is placed on the sdcard, an application
will require an explicit selinux permission to access it.
This commit adds selinux rule that allows audio_console_app
to read and write a file on the sdcard's topmost directory.

This enables playing from and saving recordings on sdcard.

Change-Id: I1441cb76d2add6c63c87fdfb0410d4813e90c280