meta-qti-auto-sepolicy: allow systemd to search wifi config dir
If a systemd service uses ProtectSystem=full option, systemd will
mount system folders read-only for the process. This requires scanning
the directories, which causes denial in the case of wifi config dir
which is owned by radio:radio. Allow scanning of wifi config dir to
let this protection be enabled.

Change-Id: I0835795b649d88fcfde835be1c55db075bcc5b30