meta-qti-auto-sepolicy: Add rules for stream update service
The stream-update service checks for the properties.txt file
in the /data/stream-update directory on target. The service
is conditionally started based on this check. The rules added
give permission to the service to access the /data dir and check
attributes of the mentioned file.

Change-Id: I56ee97d08bd55846fc12bc552c631e394398a9cb
(cherry picked from commit d74d91969ba8c1905c571847dc9cac954ab67712)